finished certification manipulating

+++ admin-api/src/main/java/com/hospital/adminapi/repository/CertificateRepository.java
@@ -1,6 +1,5 @@
 package com.hospital.adminapi.repository;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 import com.hospital.adminapi.domain.Certificate;
@@ -10,5 +9,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface CertificateRepository extends JpaRepository<Certificate, Long> {
-    Optional<Certificate> findBySerialNumber(BigInteger serialNumber);
+    Optional<Certificate> findBySerialNumber(Long serialNumber);
 }
