+++ admin-api/src/main/java/com/hospital/adminapi/repository/CertificateRevokeRepository.java
@@ -1,6 +1,5 @@
 package com.hospital.adminapi.repository;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 import com.hospital.adminapi.domain.RevokedCertificate;
@@ -9,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CertificateRevokeRepository extends JpaRepository<RevokedCertificate, Integer> {
-    Optional<RevokedCertificate> findBySerialNumber(BigInteger serialNumber);
+public interface CertificateRevokeRepository extends JpaRepository<RevokedCertificate, Long> {
+    Optional<RevokedCertificate> findBySerialNumber(Long serialNumber);
 }
