+++ admin-api/src/main/java/com/hospital/adminapi/service/certificate/dto/CertificateResponseDTO.java
@@ -1,6 +1,5 @@
 package com.hospital.adminapi.service.certificate.dto;
 
-import java.math.BigInteger;
 import java.sql.Timestamp;
 
 import com.hospital.adminapi.domain.Certificate;
@@ -9,14 +8,14 @@ import lombok.Data;
 
 @Data
 public class CertificateResponseDTO {
-    private BigInteger serialNumber;
+    private Long serialNumber;
     private Timestamp startDate;
     private Timestamp endDate;
     private String commonName;
     private boolean revoked;
 
     public CertificateResponseDTO(Certificate certificate) {
-        if(certificate != null) {
+        if (certificate != null) {
             this.serialNumber = certificate.getSerialNumber();
             this.startDate = certificate.getStartDate();
             this.endDate = certificate.getEndDate();
